refactor to decorate ClosableSequenceLoader
